refactor：重构日志展示模块
